+++ gen/c/std/deflate.c
@@ -1912,7 +1912,7 @@ inline wuffs_base__status wuffs_deflate__decoder::decode(
 
 #endif  // WUFFS_INCLUDE_GUARD__DEFLATE
 
-// !! C HEADER ENDS HERE.
+#ifdef WUFFS_IMPLEMENTATION
 
 #ifndef WUFFS_INCLUDE_GUARD__BASE_PRIVATE
 #define WUFFS_INCLUDE_GUARD__BASE_PRIVATE
@@ -4628,3 +4628,5 @@ short_read_src:
 
 #endif  // !defined(WUFFS_CONFIG__MODULES) ||
         // defined(WUFFS_CONFIG__MODULE__DEFLATE)
+
+#endif  // WUFFS_IMPLEMENTATION