+++ gen/c/std/gzip.c
@@ -2119,7 +2119,7 @@ inline wuffs_base__status wuffs_gzip__decoder::decode(
 
 #endif  // WUFFS_INCLUDE_GUARD__GZIP
 
-// !! C HEADER ENDS HERE.
+#ifdef WUFFS_IMPLEMENTATION
 
 #ifndef WUFFS_INCLUDE_GUARD__BASE_PRIVATE
 #define WUFFS_INCLUDE_GUARD__BASE_PRIVATE
@@ -3300,3 +3300,5 @@ short_read_src:
 
 #endif  // !defined(WUFFS_CONFIG__MODULES) ||
         // defined(WUFFS_CONFIG__MODULE__GZIP)
+
+#endif  // WUFFS_IMPLEMENTATION