+++ gen/c/std/lzw.c
@@ -1850,7 +1850,7 @@ inline wuffs_base__status wuffs_lzw__decoder::decode(
 
 #endif  // WUFFS_INCLUDE_GUARD__LZW
 
-// !! C HEADER ENDS HERE.
+#ifdef WUFFS_IMPLEMENTATION
 
 #ifndef WUFFS_INCLUDE_GUARD__BASE_PRIVATE
 #define WUFFS_INCLUDE_GUARD__BASE_PRIVATE
@@ -2887,3 +2887,5 @@ short_read_src:
 }
 
 #endif  // !defined(WUFFS_CONFIG__MODULES) || defined(WUFFS_CONFIG__MODULE__LZW)
+
+#endif  // WUFFS_IMPLEMENTATION